Moving some tests to EagerLoaderTest
